requirements.txt
Celery
Django==4.1.6
django-extensions==3.2.1
ipython
django_debug_toolbar==3.8.1
django-admin-rangefilter==0.9.0
django-import-export==3.1.0
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
requests
beautifulsoup4==4.12.0



#flake8
flake8
flake8-builtins
flake8-print

